requirements.txt
aiofiles>=24.1.0
bcrypt>=4.3.0
fastapi>=0.116.1
jinja2>=3.1.6
joblib>=1.5.2
numpy>=2.3.2
pandas>=2.3.2
passlib>=1.7.4
mysql-connector-python
pydantic>=2.11.7
python-jose>=3.5.0
python-multipart>=0.0.20
requests>=2.32.5
scikit-learn>=1.7.1
sqlalchemy>=2.0.43
uvicorn>=0.35.0
websockets>=15.0.1
wsproto>=1.2.0
python-dotenv>=0.21.0
